🚧♻️✅ (GURB): Improve tests structure and small fixes
