+++ som_gurb/demo/som_gurb_cups_demo.xml
@@ -5,13 +5,15 @@
             <field name="start_date">2016-01-01</field>
             <field name="gurb_id" ref="gurb_0001" />
             <field name="cups_id" ref="giscedata_cups.cups_01" />
-            <field name="beta_kw">2</field>
+            <field name="beta_kw">2.5</field>
+            <field name="extra_beta_kw">1</field>
         </record>
         <record id="gurb_cups_0002" model="som.gurb.cups">
             <field name="start_date">2019-01-22</field>
             <field name="gurb_id" ref="gurb_0001" />
             <field name="cups_id" ref="giscedata_cups.cups_02" />
-            <field name="beta_kw">1</field>
+            <field name="beta_kw">3</field>
+            <field name="extra_beta_kw">0</field>
         </record>
     </data>
 </openerp>
