+++ som_gurb/tests/__init__.py
@@ -1,2 +1,3 @@
 from tests_gurb_switching import *
 from tests_gurb_services import *
+from tests_gurb_cups import *
